+++ Sources/Telegram/Models/User.swift
@@ -3,7 +3,7 @@ import Vapor
 
 public struct User: Content {
     
-    public let id: Int
+    public let id: Int64
     public let is_bot: Bool
     public let username: String?
 }
